fastApi-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
oracledb==2.4.1
python-dotenv==1.0.1
bcrypt==4.2.0
pyjwt==2.9.0
pydantic==2.9.2